♻️ 映画取得のロジック部分をhooksとして切り出す
