+++ src/App.tsx
@@ -1,81 +1,13 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { MainLayout } from "./components/Layout/MainLayout";
 import { MovieList } from "./components/MovieList/MovieList";
 import { Search } from "./components/Search/Search";
-import type { Movie } from "./types";
-import type { TMDBApiResponse, TMDBGenre, TMDBMovie } from "./types/api/tmdb";
+import { useMovies } from "./hooks/useMovies";
 
 function App() {
-  const [movies, setMovies] = useState<Movie[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
   const [keyword, setKeyword] = useState("");
   const [year, setYear] = useState("");
-
-  useEffect(() => {
-    // キーワードが空の場合は、検索結果をクリア
-    if (!keyword.trim()) {
-      setMovies([]);
-      return;
-    }
-    // ローディング開始
-    setIsLoading(true);
-    setError(null);
-
-    const fetchMovies = async () => {
-      try {
-        // APIリクエスト用のURLパラメータを生成
-        const params = new URLSearchParams({
-          keyword: keyword.trim(),
-          ...(year && { year }),
-        });
-
-        // APIリクエスト
-        const response = await fetch(`/api/movies?${params}`);
-        if (!response.ok) {
-          throw new Error(
-            response.body?.toString() || "APIリクエストに失敗しました"
-          );
-        }
-        const data: TMDBApiResponse = await response.json();
-
-        // ジャンルIDと名前のマッピングを作成
-        const genreMap = new Map<number, string>();
-        data.genres.forEach((genre: TMDBGenre) => {
-          genreMap.set(genre.id, genre.name);
-        });
-
-        // APIレスポンスのデータをアプリケーション用に整形
-        const formattedMovies = data.moviesData.results.map(
-          (movie: TMDBMovie) => ({
-            id: movie.id,
-            title: movie.title,
-            thumbnail:
-              movie.poster_path &&
-              `https://image.tmdb.org/t/p/w500${movie.poster_path}`,
-            releaseDate: movie.release_date,
-            genres: movie.genre_ids.map(
-              (genreId) => genreMap.get(genreId) || ""
-            ),
-          })
-        );
-
-        setMovies(formattedMovies);
-      } catch (error) {
-        if (error instanceof Error) console.error(error);
-        setError(
-          error instanceof Error
-            ? error.message
-            : "予期せぬエラーが発生しました"
-        );
-        setMovies([]);
-      } finally {
-        // ローディング終了
-        setIsLoading(false);
-      }
-    };
-    fetchMovies();
-  }, [keyword, year]);
+  const { movies, isLoading, error } = useMovies(keyword, year);
 
   return (
     <MainLayout>
@@ -85,7 +17,7 @@ function App() {
         onKeywordChange={setKeyword}
         onYearChange={setYear}
       />
-      {isLoading && <div>読み込み中...</div>}
+      {isLoading && <div>Loading...</div>}
       {error && <div>{error}</div>}
       {!isLoading && !error && <MovieList movies={movies} />}
     </MainLayout>
